- Se implementa lógica para eliminar o agregar artículos a Promoción al momento de Editar una Promoción mostrando además, los artículos actuales que tiene una promoción al momento de agregar o quitar. - Se agrega validación para enviar mensaje de error si quieren eliminar una promoción que tiene artículos asignados. - Se implementa lógica para mostrar los artículos asignados que tiene una promoción en Detalles. Se mejoran las vistas: - Create. - Delete. - Details. - Edit.

+++ MVCVentas/Controllers/PromoDescuento_EController.cs
@@ -1,11 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text.Json;
+using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Caching.Memory;
 using MVCVentas.Data;
 using MVCVentas.Models;
 
@@ -15,10 +18,12 @@ namespace MVCVentas.Controllers
     public class PromoDescuento_EController : Controller
     {
         private readonly MVCVentasContext _context;
+        private readonly IMemoryCache _memoryCache;
 
-        public PromoDescuento_EController(MVCVentasContext context)
+        public PromoDescuento_EController(MVCVentasContext context, IMemoryCache memoryCache)
         {
             _context = context;
+            _memoryCache = memoryCache;
         }
 
         // GET: PromoDescuento_E
@@ -38,6 +43,8 @@ namespace MVCVentas.Controllers
 
             var vMPromoDescuento_E = await _context.VMPromoDescuento_E
                 .Include(v => v.TipoPromoDescuento)
+                .Include(v => v.ListPromoDescuento_D)
+                    .ThenInclude(v => v.Articulo)
                 .FirstOrDefaultAsync(m => m.Id_Promocion == id);
 
             if (vMPromoDescuento_E == null)
@@ -119,37 +126,87 @@ namespace MVCVentas.Controllers
         // GET: PromoDescuento_E/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            // Si el id es nulo o la entidad VMPromoDescuento_E es nula, se retorna NotFound.
             if (id == null || _context.VMPromoDescuento_E == null)
             {
                 return NotFound();
             }
 
-            var vMPromoDescuento_E = await _context.VMPromoDescuento_E.FindAsync(id);
+            // Declarar una variable de tipo VMPromoDescuento_E y asignarle el valor de la promoción con el id correspondiente.
+            var vMPromoDescuento_E = await _context.VMPromoDescuento_E
+                    .Include(v => v.ListPromoDescuento_D)
+                        .ThenInclude(v => v.Articulo)
+                    .FirstOrDefaultAsync(v => v.Id_Promocion == id);
+
+            // Asignar a la propiedad ListaArticulos de la promoción, la lista de artículos existentes.
+            vMPromoDescuento_E.ListaArticulos = _context.VMArticle
+                                .Select(a => new SelectListItem { Value = a.Id_Articulo.ToString(), Text = a.Nombre })
+                                .ToList();
+
+            // Asignar a la propiedad ArticulosSeleccionados de la promoción, la lista de artículos seleccionados.
+            vMPromoDescuento_E.ArticulosSeleccionados = vMPromoDescuento_E.ListPromoDescuento_D.Select(a => a.Id_Articulo).ToList();
+
+            // Serializar el objeto VMPromoDescuento_E y guardarlo en ViewData.
+            var options = new JsonSerializerOptions
+            {
+                ReferenceHandler = ReferenceHandler.Preserve,
+                DefaultBufferSize = 65536 // 64 kilobytes
+            };
+            var jsonModel = JsonSerializer.Serialize(vMPromoDescuento_E, options);
+            ViewData["JsonModel"] = jsonModel;
+
+            // Si la promoción es nula, se retorna NotFound.
             if (vMPromoDescuento_E == null)
             {
                 return NotFound();
             }
-            ViewData["Id_Tipo"] = new SelectList(_context.Set<VMTipoPromoDescuento>(), "Id_Tipo", "Id_Tipo", vMPromoDescuento_E.Id_Tipo);
+
+            // Se retorna la vista con la promoción correspondiente.
+            ViewData["Id_Tipo"] = new SelectList(_context.Set<VMTipoPromoDescuento>(), "Id_Tipo", "Descripcion", vMPromoDescuento_E.Id_Tipo);
             return View(vMPromoDescuento_E);
         }
 
         // POST: PromoDescuento_E/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id_Promocion,Nombre,Porcentaje,FechaInicio,FechaFin,Id_Tipo")] VMPromoDescuento_E vMPromoDescuento_E)
+        public async Task<IActionResult> Edit(int id, [Bind("Id_Promocion,Nombre,Porcentaje,FechaInicio,FechaFin,Id_Tipo,ArticulosSeleccionados")] VMPromoDescuento_E vMPromoDescuento_E)
         {
             if (id != vMPromoDescuento_E.Id_Promocion)
             {
                 return NotFound();
             }
 
+            // Si el modelo es válido, se intenta actualizar la promoción.
             if (ModelState.IsValid)
             {
                 try
                 {
                     _context.Update(vMPromoDescuento_E);
+
+                    // Si la lista de artículos seleccionados no está vacía, se recorre el foreach.
+                    if (vMPromoDescuento_E.ArticulosSeleccionados != null && vMPromoDescuento_E.ArticulosSeleccionados.Any())
+                    {
+                        var promosDescuentos_D = _context.VMPromoDescuento_D.Where(p => p.Id_Promocion == vMPromoDescuento_E.Id_Promocion).ToList();
+                        _context.RemoveRange(promosDescuentos_D);
+
+                        // Se recorre el foreach para insertar los artículos seleccionados en la base.
+                        foreach (int idArticulo in vMPromoDescuento_E.ArticulosSeleccionados)
+                        {
+                            // Recorre el primer artículo, crea una nueva instancia de VMPromoDescuento_D y se le asigna el valor de la promoción creada anteriormente y el id del artículo recorrido.
+                            VMPromoDescuento_D vMPromoDescuento_D = new VMPromoDescuento_D
+                            {
+                                Id_Promocion = vMPromoDescuento_E.Id_Promocion,
+                                Id_Articulo = idArticulo
+                            };
+                            // Se inserta en la base el objeto instanciado.
+                            _context.Add(vMPromoDescuento_D);
+                        }
+                    }
+                    else {
+                        var promosDescuentos_D = _context.VMPromoDescuento_D.Where(p => p.Id_Promocion == vMPromoDescuento_E.Id_Promocion).ToList();
+                        _context.RemoveRange(promosDescuentos_D);
+                    }
+                    // Se guardan los cambios en la base.
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
@@ -180,6 +237,7 @@ namespace MVCVentas.Controllers
             var vMPromoDescuento_E = await _context.VMPromoDescuento_E
                 .Include(v => v.TipoPromoDescuento)
                 .FirstOrDefaultAsync(m => m.Id_Promocion == id);
+
             if (vMPromoDescuento_E == null)
             {
                 return NotFound();
@@ -197,7 +255,18 @@ namespace MVCVentas.Controllers
             {
                 return Problem("Entity set 'MVCVentasContext.VMPromoDescuento_E'  is null.");
             }
-            var vMPromoDescuento_E = await _context.VMPromoDescuento_E.FindAsync(id);
+
+            var vMPromoDescuento_E = await _context.VMPromoDescuento_E
+                .Include(v => v.ListPromoDescuento_D)
+                    .ThenInclude(v => v.Articulo)
+                .FirstOrDefaultAsync(v => v.Id_Promocion == id);
+            
+            if (vMPromoDescuento_E.ListPromoDescuento_D != null && vMPromoDescuento_E.ListPromoDescuento_D.Any())
+            {
+                TempData["MensajeError"] = "No se puede eliminar la promoción, ya que tiene artículos asociados.";
+                return RedirectToAction("Delete");
+            }
+
             if (vMPromoDescuento_E != null)
             {
                 _context.VMPromoDescuento_E.Remove(vMPromoDescuento_E);
